requirements.txt
numpy==1.24.4
scipy==1.15.3
scikit-learn==1.3.2
h5py==3.10.0
Pillow==10.4.0
opencv-python==4.8.1.78
scikit-image==0.21.0
imageio==2.37.0
kornia==0.7.3
plyfile==1.0.3
PyMCubes==0.1.4
taichi==1.5.0
warp-lang==0.10.1
transformers==4.57.1
einops==0.8.0
jaxtyping==0.2.19
causal-conv1d==1.4.0
mamba-ssm==2.2.2
xformers==0.0.27
matplotlib==3.7.5
lpips==0.1.4
kiui==0.2.18
tensorboard==2.20.0
wandb==0.19.1
easydict==1.13
PyYAML==6.0.2
tqdm==4.67.1
termcolor==2.4.0
packaging==24.2
typing_extensions==4.15.0